refactor: extract formatResetTime to shared utils
